pyproject.toml
[tool.poetry]
name = "yaic"
version = "0.1.0"
description = "Yet Another Image Classifier"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "yaic" }]

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.poetry.scripts]
yaic = "yaic.main:main"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
